+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [v0.3.4](https://github.com/jinhucheung/letscertbot/tree/v0.3.4) (2020-02-10)
+
+[Full Changelog](https://github.com/jinhucheung/letscertbot/compare/v0.3.3..v0.3.4)
+
+**Implemented enhancements:**
+
+- Add after_hook to Deploy script [\#18](https://github.com/jinhucheung/letscertbot/pull/18)
+
 ## [v0.3.3](https://github.com/jinhucheung/letscertbot/tree/v0.3.3) (2020-02-06)
 
 [Full Changelog](https://github.com/jinhucheung/letscertbot/compare/v0.3.2..v0.3.3)